refactor(main): use css for number of months to display
