+++ lib/utils.js
@@ -87,3 +87,9 @@ export function getToday () {
     now.getHours()
   ]
 }
+
+export function parseNumber (str, fallback) {
+  const num = parseInt(str, 10)
+  const result = Number.isNaN(num) ? fallback : num
+  return result
+}
